+++ org.idempiere.webservices/WEB-INF/src/org/idempiere/webservices/AbstractService.java
@@ -34,13 +34,11 @@ import org.adempiere.exceptions.AdempiereException;
 import org.apache.commons.codec.binary.Base64;
 import org.compiere.model.Lookup;
 import org.compiere.model.MUser;
-import org.compiere.model.MWebService;
-import org.compiere.model.MWebServiceType;
 import org.compiere.model.PO;
 import org.compiere.model.POInfo;
 import org.compiere.model.Query;
-import org.compiere.model.X_WS_WebServiceMethod;
-import org.compiere.model.X_WS_WebServiceTypeAccess;
+import org.idempiere.webservices.model.X_WS_WebServiceMethod;
+import org.idempiere.webservices.model.X_WS_WebServiceTypeAccess;
 import org.compiere.util.CCache;
 import org.compiere.util.DB;
 import org.compiere.util.Env;
@@ -57,6 +55,8 @@ import org.idempiere.adInterface.x10.StandardResponseDocument;
 import org.idempiere.adinterface.CompiereService;
 import org.idempiere.cache.ImmutablePOCache;
 import org.idempiere.webservices.fault.IdempiereServiceFault;
+import org.idempiere.webservices.model.MWebService;
+import org.idempiere.webservices.model.MWebServiceType;
