+++ org.idempiere.webservices/WEB-INF/src/org/idempiere/webservices/model/I_WS_WebServiceFieldOutput.java
@@ -14,24 +14,24 @@
  * ComPiere, Inc., 2620 Augustine Dr. #245, Santa Clara, CA 95054, USA        *
  * or via info@compiere.org or http://www.compiere.org/license.html           *
  *****************************************************************************/
-package org.compiere.model;
+package org.idempiere.webservices.model;
 
 import java.math.BigDecimal;
 import java.sql.Timestamp;
 import org.compiere.util.KeyNamePair;
 
-/** Generated Interface for WS_WebServiceTypeAccess
+/** Generated Interface for WS_WebServiceFieldOutput
  *  @author iDempiere (generated) 
  *  @version Development 9.0
  */
-public interface I_WS_WebServiceTypeAccess 
+public interface I_WS_WebServiceFieldOutput 
 {
 
-    /** TableName=WS_WebServiceTypeAccess */
-    public static final String Table_Name = "WS_WebServiceTypeAccess";
+    /** TableName=WS_WebServiceFieldOutput */
+    public static final String Table_Name = "WS_WebServiceFieldOutput";
 
-    /** AD_Table_ID=53168 */
-    public static final int Table_ID = 53168;
+    /** AD_Table_ID=53167 */
+    public static final int Table_ID = 53167;
 
     KeyNamePair Model = new KeyNamePair(Table_ID, Table_Name);
 
@@ -49,6 +49,21 @@ public interface I_WS_WebServiceTypeAccess
 	  */
 	public int getAD_Client_ID();
 
+    /** Column name AD_Column_ID */
+    public static final String COLUMNNAME_AD_Column_ID = "AD_Column_ID";
+
+	/** Set Column.
+	  * Column in the table
+	  */
+	public void setAD_Column_ID (int AD_Column_ID);
+
+	/** Get Column.
+	  * Column in the table
+	  */
+	public int getAD_Column_ID();
+
+	public org.compiere.model.I_AD_Column getAD_Column() throws RuntimeException;
+
     /** Column name AD_Org_ID */
     public static final String COLUMNNAME_AD_Org_ID = "AD_Org_ID";
 
@@ -61,21 +76,6 @@ public interface I_WS_WebServiceTypeAccess
 	  * Organizational entity within client
 	  */
 	public int getAD_Org_ID();
-
-    /** Column name AD_Role_ID */
-    public static final String COLUMNNAME_AD_Role_ID = "AD_Role_ID";
-
-	/** Set Role.
-	  * Responsibility Role
-	  */
-	public void setAD_Role_ID (int AD_Role_ID);
-
-	/** Get Role.
-	  * Responsibility Role
-	  */
-	public int getAD_Role_ID();
-
-	public org.compiere.model.I_AD_Role getAD_Role() throws RuntimeException;
 
     /** Column name Created */
     public static final String COLUMNNAME_Created = "Created";
@@ -106,19 +106,6 @@ public interface I_WS_WebServiceTypeAccess
 	  */
 	public boolean isActive();
 
-    /** Column name IsReadWrite */
-    public static final String COLUMNNAME_IsReadWrite = "IsReadWrite";
-
-	/** Set Read Write.
-	  * Field is read / write
-	  */
-	public void setIsReadWrite (boolean IsReadWrite);
-
-	/** Get Read Write.
-	  * Field is read / write
-	  */
-	public boolean isReadWrite();
-
     /** Column name Updated */
     public static final String COLUMNNAME_Updated = "Updated";
 
@@ -135,14 +122,23 @@ public interface I_WS_WebServiceTypeAccess
 	  */
 	public int getUpdatedBy();
 
-    /** Column name WS_WebServiceTypeAccess_UU */
-    public static final String COLUMNNAME_WS_WebServiceTypeAccess_UU = "WS_WebServiceTypeAccess_UU";
+    /** Column name WS_WebServiceFieldOutput_ID */
+    public static final String COLUMNNAME_WS_WebServiceFieldOutput_ID = "WS_WebServiceFieldOutput_ID";
 
-	/** Set WS_WebServiceTypeAccess_UU	  */
-	public void setWS_WebServiceTypeAccess_UU (String WS_WebServiceTypeAccess_UU);
+	/** Set Web Service Field Output	  */
+	public void setWS_WebServiceFieldOutput_ID (int WS_WebServiceFieldOutput_ID);
 
-	/** Get WS_WebServiceTypeAccess_UU	  */
-	public String getWS_WebServiceTypeAccess_UU();
+	/** Get Web Service Field Output	  */
+	public int getWS_WebServiceFieldOutput_ID();
+
+    /** Column name WS_WebServiceFieldOutput_UU */
+    public static final String COLUMNNAME_WS_WebServiceFieldOutput_UU = "WS_WebServiceFieldOutput_UU";
+
+	/** Set WS_WebServiceFieldOutput_UU	  */
+	public void setWS_WebServiceFieldOutput_UU (String WS_WebServiceFieldOutput_UU);
+
+	/** Get WS_WebServiceFieldOutput_UU	  */
+	public String getWS_WebServiceFieldOutput_UU();
 
     /** Column name WS_WebServiceType_ID */
     public static final String COLUMNNAME_WS_WebServiceType_ID = "WS_WebServiceType_ID";
@@ -153,5 +149,5 @@ public interface I_WS_WebServiceTypeAccess
 	/** Get Web Service Type	  */
 	public int getWS_WebServiceType_ID();
 
-	public org.compiere.model.I_WS_WebServiceType getWS_WebServiceType() throws RuntimeException;
+	public I_WS_WebServiceType getWS_WebServiceType() throws RuntimeException;
 }
