+++ org.idempiere.webservices/WEB-INF/src/org/idempiere/webservices/model/I_WS_WebServiceMethod.java
@@ -14,30 +14,30 @@
  * ComPiere, Inc., 2620 Augustine Dr. #245, Santa Clara, CA 95054, USA        *
  * or via info@compiere.org or http://www.compiere.org/license.html           *
  *****************************************************************************/
-package org.compiere.model;
+package org.idempiere.webservices.model;
 
 import java.math.BigDecimal;
 import java.sql.Timestamp;
 import org.compiere.util.KeyNamePair;
 
-/** Generated Interface for WS_WebService_Para
+/** Generated Interface for WS_WebServiceMethod
  *  @author iDempiere (generated) 
  *  @version Development 9.0
  */
-public interface I_WS_WebService_Para 
+public interface I_WS_WebServiceMethod 
 {
 
-    /** TableName=WS_WebService_Para */
-    public static final String Table_Name = "WS_WebService_Para";
+    /** TableName=WS_WebServiceMethod */
+    public static final String Table_Name = "WS_WebServiceMethod";
 
-    /** AD_Table_ID=53165 */
-    public static final int Table_ID = 53165;
+    /** AD_Table_ID=53163 */
+    public static final int Table_ID = 53163;
 
     KeyNamePair Model = new KeyNamePair(Table_ID, Table_Name);
 
-    /** AccessLevel = 6 - System - Client 
+    /** AccessLevel = 4 - System 
      */
-    BigDecimal accessLevel = BigDecimal.valueOf(6);
+    BigDecimal accessLevel = BigDecimal.valueOf(4);
 
     /** Load Meta Data */
 
@@ -62,19 +62,6 @@ public interface I_WS_WebService_Para
 	  */
 	public int getAD_Org_ID();
 
-    /** Column name ConstantValue */
-    public static final String COLUMNNAME_ConstantValue = "ConstantValue";
-
-	/** Set Constant Value.
-	  * Constant value
-	  */
-	public void setConstantValue (String ConstantValue);
-
-	/** Get Constant Value.
-	  * Constant value
-	  */
-	public String getConstantValue();
-
     /** Column name Created */
     public static final String COLUMNNAME_Created = "Created";
 
@@ -91,6 +78,32 @@ public interface I_WS_WebService_Para
 	  */
 	public int getCreatedBy();
 
+    /** Column name Description */
+    public static final String COLUMNNAME_Description = "Description";
+
+	/** Set Description.
+	  * Optional short description of the record
+	  */
+	public void setDescription (String Description);
+
+	/** Get Description.
+	  * Optional short description of the record
+	  */
+	public String getDescription();
+
+    /** Column name Help */
+    public static final String COLUMNNAME_Help = "Help";
+
+	/** Set Comment/Help.
+	  * Comment or Hint
+	  */
+	public void setHelp (String Help);
+
+	/** Get Comment/Help.
+	  * Comment or Hint
+	  */
+	public String getHelp();
+
     /** Column name IsActive */
     public static final String COLUMNNAME_IsActive = "IsActive";
 
@@ -104,23 +117,18 @@ public interface I_WS_WebService_Para
 	  */
 	public boolean isActive();
 
-    /** Column name ParameterName */
-    public static final String COLUMNNAME_ParameterName = "ParameterName";
+    /** Column name Name */
+    public static final String COLUMNNAME_Name = "Name";
 
-	/** Set Parameter Name	  */
-	public void setParameterName (String ParameterName);
+	/** Set Name.
+	  * Alphanumeric identifier of the entity
+	  */
+	public void setName (String Name);
 
-	/** Get Parameter Name	  */
-	public String getParameterName();
-
-    /** Column name ParameterType */
-    public static final String COLUMNNAME_ParameterType = "ParameterType";
-
-	/** Set Parameter Type	  */
-	public void setParameterType (String ParameterType);
-
-	/** Get Parameter Type	  */
-	public String getParameterType();
+	/** Get Name.
+	  * Alphanumeric identifier of the entity
+	  */
+	public String getName();
 
     /** Column name Updated */
     public static final String COLUMNNAME_Updated = "Updated";
@@ -138,32 +146,45 @@ public interface I_WS_WebService_Para
 	  */
 	public int getUpdatedBy();
 
-    /** Column name WS_WebService_Para_ID */
-    public static final String COLUMNNAME_WS_WebService_Para_ID = "WS_WebService_Para_ID";
+    /** Column name Value */
+    public static final String COLUMNNAME_Value = "Value";
 
-	/** Set Web Service Parameters	  */
-	public void setWS_WebService_Para_ID (int WS_WebService_Para_ID);
+	/** Set Search Key.
+	  * Search key for the record in the format required - must be unique
+	  */
+	public void setValue (String Value);
 
-	/** Get Web Service Parameters	  */
-	public int getWS_WebService_Para_ID();
+	/** Get Search Key.
+	  * Search key for the record in the format required - must be unique
+	  */
+	public String getValue();
 
-    /** Column name WS_WebService_Para_UU */
-    public static final String COLUMNNAME_WS_WebService_Para_UU = "WS_WebService_Para_UU";
+    /** Column name WS_WebService_ID */
+    public static final String COLUMNNAME_WS_WebService_ID = "WS_WebService_ID";
 
-	/** Set WS_WebService_Para_UU	  */
-	public void setWS_WebService_Para_UU (String WS_WebService_Para_UU);
+	/** Set Web Service	  */
+	public void setWS_WebService_ID (int WS_WebService_ID);
 
-	/** Get WS_WebService_Para_UU	  */
-	public String getWS_WebService_Para_UU();
+	/** Get Web Service	  */
+	public int getWS_WebService_ID();
 
-    /** Column name WS_WebServiceType_ID */
-    public static final String COLUMNNAME_WS_WebServiceType_ID = "WS_WebServiceType_ID";
+	public I_WS_WebService getWS_WebService() throws RuntimeException;
 
-	/** Set Web Service Type	  */
-	public void setWS_WebServiceType_ID (int WS_WebServiceType_ID);
+    /** Column name WS_WebServiceMethod_ID */
+    public static final String COLUMNNAME_WS_WebServiceMethod_ID = "WS_WebServiceMethod_ID";
 
-	/** Get Web Service Type	  */
-	public int getWS_WebServiceType_ID();
+	/** Set Web Service Method	  */
+	public void setWS_WebServiceMethod_ID (int WS_WebServiceMethod_ID);
 
-	public org.compiere.model.I_WS_WebServiceType getWS_WebServiceType() throws RuntimeException;
+	/** Get Web Service Method	  */
+	public int getWS_WebServiceMethod_ID();
+
+    /** Column name WS_WebServiceMethod_UU */
+    public static final String COLUMNNAME_WS_WebServiceMethod_UU = "WS_WebServiceMethod_UU";
+
+	/** Set WS_WebServiceMethod_UU	  */
+	public void setWS_WebServiceMethod_UU (String WS_WebServiceMethod_UU);
+
+	/** Get WS_WebServiceMethod_UU	  */
+	public String getWS_WebServiceMethod_UU();
 }
