+++ org.idempiere.webservices/WEB-INF/src/org/idempiere/webservices/model/MWebService.java
@@ -27,7 +27,7 @@
 * - GlobalQSS (http://www.globalqss.com)                              *
 ***********************************************************************/
 
-package org.compiere.model;
+package org.idempiere.webservices.model;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
