+++ org.idempiere.webservices/WEB-INF/src/org/idempiere/webservices/model/MWebServiceMethod.java
@@ -20,27 +20,82 @@
 * MA 02110-1301, USA.                                                 *
 *                                                                     *
 * Contributors:                                                       *
-* - Carlos Ruiz - globalqss                                           *
+* - Trek Global Corporation                                           *
+* - Heng Sin Low                                                      *
 **********************************************************************/
+package org.idempiere.webservices.model;
 
-package org.compiere.model;
+import java.sql.ResultSet;
+import java.util.Properties;
 
-import org.adempiere.base.AnnotationBasedModelFactory;
-import org.adempiere.base.IModelFactory;
-import org.osgi.service.component.annotations.Component;
+import org.compiere.util.Env;
+import org.idempiere.cache.ImmutablePOSupport;
 
 /**
- * 
- * @author globalqss
  * @author hengsin
  *
  */
-@Component(immediate = true, service = IModelFactory.class, property = "service.ranking:Integer=1")
-public class WS_ModelFactory extends AnnotationBasedModelFactory {
+@org.adempiere.base.Model(table = X_WS_WebServiceMethod.Table_Name)
+public class MWebServiceMethod extends X_WS_WebServiceMethod implements ImmutablePOSupport {
 
-	@Override
-	protected String[] getPackages() {
-		return new String[] {MWebService.class.getPackageName()};
+	/**
+	 * 
+	 */
+	private static final long serialVersionUID = 6822623209157500849L;
+
+	/**
+	 * @param ctx
+	 * @param WS_WebServiceMethod_ID
+	 * @param trxName
+	 */
+	public MWebServiceMethod(Properties ctx, int WS_WebServiceMethod_ID, String trxName) {
+		super(ctx, WS_WebServiceMethod_ID, trxName);
+	}
+
+	/**
+	 * @param ctx
+	 * @param rs
+	 * @param trxName
+	 */
+	public MWebServiceMethod(Properties ctx, ResultSet rs, String trxName) {
+		super(ctx, rs, trxName);
+	}
+
+	/**
+	 * 
+	 * @param copy
+	 */
+	public MWebServiceMethod(MWebServiceMethod copy) {
+		this(Env.getCtx(), copy);
+	}
+
+	/**
+	 * 
+	 * @param ctx
+	 * @param copy
+	 */
+	public MWebServiceMethod(Properties ctx, MWebServiceMethod copy) {
+		this(ctx, copy, (String) null);
+	}
+
+	/**
+	 * 
+	 * @param ctx
+	 * @param copy
+	 * @param trxName
+	 */
+	public MWebServiceMethod(Properties ctx, MWebServiceMethod copy, String trxName) {
+		this(ctx, 0, trxName);
+		copyPO(copy);
 	}
 	
+	@Override
+	public MWebServiceMethod markImmutable() {
+		if (is_Immutable())
+			return this;
+
+		makeImmutable();
+		return this;
+	}
+
 }
