+++ org.idempiere.webservices/WEB-INF/src/org/idempiere/webservices/model/X_WS_WebServiceFieldOutput.java
@@ -15,10 +15,15 @@
  * or via info@compiere.org or http://www.compiere.org/license.html           *
  *****************************************************************************/
 /** Generated Model - DO NOT CHANGE */
-package org.compiere.model;
+package org.idempiere.webservices.model;
 
 import java.sql.ResultSet;
 import java.util.Properties;
+
+import org.compiere.model.I_Persistent;
+import org.compiere.model.MTable;
+import org.compiere.model.PO;
+import org.compiere.model.POInfo;
 import org.compiere.util.KeyNamePair;
 
 /** Generated Model for WS_WebServiceFieldOutput
@@ -143,9 +148,9 @@ public class X_WS_WebServiceFieldOutput extends PO implements I_WS_WebServiceFie
 		return (String)get_Value(COLUMNNAME_WS_WebServiceFieldOutput_UU);
 	}
 
-	public org.compiere.model.I_WS_WebServiceType getWS_WebServiceType() throws RuntimeException
+	public I_WS_WebServiceType getWS_WebServiceType() throws RuntimeException
 	{
-		return (org.compiere.model.I_WS_WebServiceType)MTable.get(getCtx(), org.compiere.model.I_WS_WebServiceType.Table_ID)
+		return (I_WS_WebServiceType)MTable.get(getCtx(), I_WS_WebServiceType.Table_ID)
 			.getPO(getWS_WebServiceType_ID(), get_TrxName());
 	}
