+++ org.idempiere.webservices/WEB-INF/src/org/idempiere/webservices/model/X_WS_WebServiceType.java
@@ -15,17 +15,22 @@
  * or via info@compiere.org or http://www.compiere.org/license.html           *
  *****************************************************************************/
 /** Generated Model - DO NOT CHANGE */
-package org.compiere.model;
+package org.idempiere.webservices.model;
 
 import java.sql.ResultSet;
 import java.util.Properties;
+
+import org.compiere.model.I_Persistent;
+import org.compiere.model.MTable;
+import org.compiere.model.PO;
+import org.compiere.model.POInfo;
 import org.compiere.util.KeyNamePair;
 
-/** Generated Model for WS_WebService
+/** Generated Model for WS_WebServiceType
  *  @author iDempiere (generated) 
  *  @version Development 9.0 - $Id$ */
-@org.adempiere.base.Model(table="WS_WebService")
-public class X_WS_WebService extends PO implements I_WS_WebService, I_Persistent 
+@org.adempiere.base.Model(table="WS_WebServiceType")
+public class X_WS_WebServiceType extends PO implements I_WS_WebServiceType, I_Persistent 
 {
 
 	/**
@@ -34,25 +39,27 @@ public class X_WS_WebService extends PO implements I_WS_WebService, I_Persistent
 	private static final long serialVersionUID = 20210917L;
 
     /** Standard Constructor */
-    public X_WS_WebService (Properties ctx, int WS_WebService_ID, String trxName)
+    public X_WS_WebServiceType (Properties ctx, int WS_WebServiceType_ID, String trxName)
     {
-      super (ctx, WS_WebService_ID, trxName);
-      /** if (WS_WebService_ID == 0)
+      super (ctx, WS_WebServiceType_ID, trxName);
+      /** if (WS_WebServiceType_ID == 0)
         {
 			setName (null);
 			setValue (null);
 			setWS_WebService_ID (0);
+			setWS_WebServiceMethod_ID (0);
+			setWS_WebServiceType_ID (0);
         } */
     }
 
     /** Load Constructor */
-    public X_WS_WebService (Properties ctx, ResultSet rs, String trxName)
+    public X_WS_WebServiceType (Properties ctx, ResultSet rs, String trxName)
     {
       super (ctx, rs, trxName);
     }
 
     /** AccessLevel
-      * @return 4 - System 
+      * @return 6 - System - Client 
       */
     protected int get_AccessLevel()
     {
@@ -68,10 +75,39 @@ public class X_WS_WebService extends PO implements I_WS_WebService, I_Persistent
 
     public String toString()
     {
-      StringBuilder sb = new StringBuilder ("X_WS_WebService[")
+      StringBuilder sb = new StringBuilder ("X_WS_WebServiceType[")
         .append(get_ID()).append(",Name=").append(getName()).append("]");
       return sb.toString();
     }
+
+	public org.compiere.model.I_AD_Table getAD_Table() throws RuntimeException
+	{
+		return (org.compiere.model.I_AD_Table)MTable.get(getCtx(), org.compiere.model.I_AD_Table.Table_ID)
+			.getPO(getAD_Table_ID(), get_TrxName());
+	}
+
+	/** Set Table.
+		@param AD_Table_ID 
+		Database Table information
+	  */
+	public void setAD_Table_ID (int AD_Table_ID)
+	{
+		if (AD_Table_ID < 1) 
+			set_Value (COLUMNNAME_AD_Table_ID, null);
+		else 
+			set_Value (COLUMNNAME_AD_Table_ID, Integer.valueOf(AD_Table_ID));
+	}
+
+	/** Get Table.
+		@return Database Table information
+	  */
+	public int getAD_Table_ID () 
+	{
+		Integer ii = (Integer)get_Value(COLUMNNAME_AD_Table_ID);
+		if (ii == null)
+			 return 0;
+		return ii.intValue();
+	}
 
 	/** Set Description.
 		@param Description 
@@ -105,6 +141,20 @@ public class X_WS_WebService extends PO implements I_WS_WebService, I_Persistent
 	public String getHelp () 
 	{
 		return (String)get_Value(COLUMNNAME_Help);
+	}
+
+	/** Set Create Default Parameters.
+		@param InsertParameters Create Default Parameters	  */
+	public void setInsertParameters (String InsertParameters)
+	{
+		set_Value (COLUMNNAME_InsertParameters, InsertParameters);
+	}
+
+	/** Get Create Default Parameters.
+		@return Create Default Parameters	  */
+	public String getInsertParameters () 
+	{
+		return (String)get_Value(COLUMNNAME_InsertParameters);
 	}
 
 	/** Set Name.
@@ -149,14 +199,20 @@ public class X_WS_WebService extends PO implements I_WS_WebService, I_Persistent
 		return (String)get_Value(COLUMNNAME_Value);
 	}
 
+	public I_WS_WebService getWS_WebService() throws RuntimeException
+	{
+		return (I_WS_WebService)MTable.get(getCtx(), I_WS_WebService.Table_ID)
+			.getPO(getWS_WebService_ID(), get_TrxName());
+	}
+
 	/** Set Web Service.
 		@param WS_WebService_ID Web Service	  */
 	public void setWS_WebService_ID (int WS_WebService_ID)
 	{
 		if (WS_WebService_ID < 1) 
-			set_ValueNoCheck (COLUMNNAME_WS_WebService_ID, null);
+			set_Value (COLUMNNAME_WS_WebService_ID, null);
 		else 
-			set_ValueNoCheck (COLUMNNAME_WS_WebService_ID, Integer.valueOf(WS_WebService_ID));
+			set_Value (COLUMNNAME_WS_WebService_ID, Integer.valueOf(WS_WebService_ID));
 	}
 
 	/** Get Web Service.
@@ -169,17 +225,63 @@ public class X_WS_WebService extends PO implements I_WS_WebService, I_Persistent
 		return ii.intValue();
 	}
 
-	/** Set WS_WebService_UU.
-		@param WS_WebService_UU WS_WebService_UU	  */
-	public void setWS_WebService_UU (String WS_WebService_UU)
+	public I_WS_WebServiceMethod getWS_WebServiceMethod() throws RuntimeException
 	{
-		set_Value (COLUMNNAME_WS_WebService_UU, WS_WebService_UU);
+		return (I_WS_WebServiceMethod)MTable.get(getCtx(), I_WS_WebServiceMethod.Table_ID)
+			.getPO(getWS_WebServiceMethod_ID(), get_TrxName());
 	}
 
-	/** Get WS_WebService_UU.
-		@return WS_WebService_UU	  */
-	public String getWS_WebService_UU () 
+	/** Set Web Service Method.
+		@param WS_WebServiceMethod_ID Web Service Method	  */
+	public void setWS_WebServiceMethod_ID (int WS_WebServiceMethod_ID)
 	{
-		return (String)get_Value(COLUMNNAME_WS_WebService_UU);
+		if (WS_WebServiceMethod_ID < 1) 
+			set_Value (COLUMNNAME_WS_WebServiceMethod_ID, null);
+		else 
+			set_Value (COLUMNNAME_WS_WebServiceMethod_ID, Integer.valueOf(WS_WebServiceMethod_ID));
+	}
+
+	/** Get Web Service Method.
+		@return Web Service Method	  */
+	public int getWS_WebServiceMethod_ID () 
+	{
+		Integer ii = (Integer)get_Value(COLUMNNAME_WS_WebServiceMethod_ID);
+		if (ii == null)
+			 return 0;
+		return ii.intValue();
+	}
+
+	/** Set Web Service Type.
+		@param WS_WebServiceType_ID Web Service Type	  */
+	public void setWS_WebServiceType_ID (int WS_WebServiceType_ID)
+	{
+		if (WS_WebServiceType_ID < 1) 
+			set_ValueNoCheck (COLUMNNAME_WS_WebServiceType_ID, null);
+		else 
+			set_ValueNoCheck (COLUMNNAME_WS_WebServiceType_ID, Integer.valueOf(WS_WebServiceType_ID));
+	}
+
+	/** Get Web Service Type.
+		@return Web Service Type	  */
+	public int getWS_WebServiceType_ID () 
+	{
+		Integer ii = (Integer)get_Value(COLUMNNAME_WS_WebServiceType_ID);
+		if (ii == null)
+			 return 0;
+		return ii.intValue();
+	}
+
+	/** Set WS_WebServiceType_UU.
+		@param WS_WebServiceType_UU WS_WebServiceType_UU	  */
+	public void setWS_WebServiceType_UU (String WS_WebServiceType_UU)
+	{
+		set_Value (COLUMNNAME_WS_WebServiceType_UU, WS_WebServiceType_UU);
+	}
+
+	/** Get WS_WebServiceType_UU.
+		@return WS_WebServiceType_UU	  */
+	public String getWS_WebServiceType_UU () 
+	{
+		return (String)get_Value(COLUMNNAME_WS_WebServiceType_UU);
 	}
 }
